feat(core/biblio): teach biblio about biblio-DB
